SELC-1623 fixed already onboarded api
SELC-1623 - fixed VERIFY_ONBOARDING api call
