+++ src/components/steps/StepBillingData.tsx
@@ -6,7 +6,7 @@ import { Grid, TextField, Typography, useTheme, Paper } from '@mui/material';
 import { useFormik } from 'formik';
 import { useTranslation } from 'react-i18next';
 import Checkbox from '@mui/material/Checkbox';
-import { useContext, useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import {
   BillingData,
   InstitutionType,
@@ -17,9 +17,7 @@ import {
 } from '../../../types';
 import { OnboardingStepActions } from '../OnboardingStepActions';
 import { useHistoryState } from '../useHistoryState';
-import { UserContext } from '../../lib/context';
 import { MessageNoAction } from '../MessageNoAction';
-import { verifyOnboarding } from '../verifyOnboarding';
 
 const CustomTextField = styled(TextField)({
   '.MuiInputLabel-asterisk': {
@@ -57,10 +55,10 @@ type Props = StepperStepComponentProps & {
   subtitle: string;
   externalInstitutionId: string;
   origin: string;
-  setExternalInstitutionId: React.Dispatch<React.SetStateAction<string>>;
   productId?: string;
   selectedParty?: Party;
   selectedProduct?: Product | null;
+  outcome?: RequestOutcomeMessage | null;
 };
 
 export default function StepBillingData({
@@ -71,17 +69,12 @@ export default function StepBillingData({
   institutionType,
   externalInstitutionId,
   origin,
-  setExternalInstitutionId,
-  productId,
-  selectedParty,
-  selectedProduct,
+  outcome,
 }: Props) {
   const requiredError = 'Required';
   const ipa = origin === 'IPA';
   const disabledField = ipa || institutionType !== 'PSP';
   const isPSP = institutionType === 'PSP';
-  const { setRequiredLogin } = useContext(UserContext);
-  const [outcome, setOutcome] = useState<RequestOutcomeMessage | null>();
 
   const { t } = useTranslation();
   const theme = useTheme();
@@ -110,23 +103,7 @@ export default function StepBillingData({
   const saveHistoryState = () => {
     setStepHistoryStateHistory(stepHistoryState);
   };
-  const validateForward = () => {
-    if (isPSP) {
-      // TODO: fix when party registry proxy will return externalInstitutionId
-      setExternalInstitutionId(formik.values.taxCode);
-      void verifyOnboarding({
-        selectedParty,
-        selectedProduct,
-        externalInstitutionId,
-        productId,
-        setOutcome,
-        onForwardAction,
-        setRequiredLogin,
-      });
-    } else {
-      onForwardAction();
-    }
-  };
+
   const onForwardAction = () => {
     saveHistoryState();
     forward({
@@ -560,7 +537,7 @@ export default function StepBillingData({
           <OnboardingStepActions
             back={{ action: onBackAction, label: t('stepBillingData.backLabel'), disabled: false }}
             forward={{
-              action: validateForward,
+              action: onForwardAction,
               label: t('stepBillingData.confirmLabel'),
               disabled: !formik.isValid,
             }}
